added bottom navbar
moved lab js files to components folder
installed react-navigation/bottom-tabs
minor cleaning of the code
