+++ mukhinDmitry/App.js
@@ -1,57 +1,20 @@
 import React, { useState } from 'react';
-import {
-  SafeAreaView,
-  StyleSheet,
-  Text,
-  View,
-  Button,
-} from 'react-native';
+import rnLab1 from './components/labs/Lab1';
+import rnLab2 from './components/labs/Lab2';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-const colorStep = (bgColor) => {
-  let bgStr = bgColor.substr(1,2)
-  bgStr = (bgStr * 1 + 5).toString()
-  bgStr = bgStr + bgStr + bgStr
-  return "#" + bgStr
-}
+const Tab = createBottomTabNavigator();
 
 const rnApp = () => {
-  const [bgColor, setBgColor] = useState('#000000');
   return (
-    <SafeAreaView>
-      <View>
-        <Text style = {styles.rnTitle}>
-          Lab 1. Color lightener
-        </Text>
-        <View style={[styles.rnBox, {backgroundColor: bgColor}]}>
-
-        </View>
-        <Button 
-          style = {styles.rnButton}
-          onPress={() => setBgColor(colorStep(bgColor))}
-          title="Try me"
-          color={bgColor}
-          accessibilityLabel="Just give it a try"
-        />
-      </View>
-    </SafeAreaView>
+    <NavigationContainer>
+      <Tab.Navigator initialRouteName="Lab 2">
+        <Tab.Screen name="Lab 1" component={rnLab1} />
+        <Tab.Screen name="Lab 2" component={rnLab2} />
+      </Tab.Navigator>
+    </NavigationContainer>
   );
 };
-
-const styles = StyleSheet.create({
-  rnButton: {
-    marginTop: 64,
-    paddingHorizontal: 24,
-  },
-  rnTitle: {
-    marginTop: 32,
-    fontSize: 24,
-    fontWeight: '600',
-    textAlign: 'center',
-  },
-  rnBox: {
-    height: 100,
-    margin: 50,
-  }
-});
 
 export default rnApp;
